% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decompGoldstein.R
\name{decompGoldstein}
\alias{decompGoldstein}
\title{Decompose two level by Goldstein method}
\usage{
decompGoldstein(allIndicators, l1Var, l2Var, df)
}
\arguments{
\item{allIndicators}{character vector of indicator names}

\item{l1Var}{string}

\item{l2Var}{string}

\item{df}{dataframe}
}
\description{
Desc:
}
\details{
Details to be written
}
