% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getVCE.R
\name{getVCE}
\alias{getVCE}
\title{Extract the variance parameter from nlme object}
\usage{
getVCE(twoLvlnlmeObj, names)
}
\arguments{
\item{twoLvlnlmeObj}{nlme model obj}

\item{names}{}
}
\description{
Desc:
}
\details{
Details to be written
}
