% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcfa.input.R
\name{mcfa.input}
\alias{mcfa.input}
\title{Muthen Decomp}
\usage{
mcfa.input(l1Var, l2Var, dat)
}
\arguments{
\item{dat}{dataframe}
}
\description{
Takes a DF and clustering variables and returns a list including decomposed covariance matrices
}
\details{
Details to be written
}
