% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlcfaMIIV.R
\name{mlcfaMIIV}
\alias{mlcfaMIIV}
\title{MIIV-2SLS estimation for multilevel cfa}
\usage{
mlcfaMIIV(withinModel, betweenModel, estimator = "muthen", allIndicators,
  l1Var, l2Var, df, var.cov = FALSE)
}
\arguments{
\item{withinModel}{Model statement using Lavaan/MIIVsem syntax}

\item{betweenModel}{Model statement using Lavaan/MIIVsem syntax}

\item{estimator}{"muthen" or "goldstein"}

\item{allIndicators}{character vector of the names of indicators}
}
\description{
Description
}
\details{
Details:
}
